move dts-downloader d.ts -> rs transform logic to build script
