+++ dts-downloader/src/lib.rs
@@ -1,18 +1,12 @@
 use std::env;
-use std::fs::File;
-use std::io::{self, BufWriter, Write};
 use std::path::PathBuf;
-use std::process::Command;
 
 use anyhow::Result;
-
-use github_webhook_type_generator::dts2rs;
 
 #[derive(Default)]
 pub struct Opt {
     pub version: Version,
     pub out_path_ts: OutPathTs,
-    pub out_path_rs: OutPathRs,
 }
 
 pub struct Version(pub String);
@@ -23,7 +17,7 @@ impl Default for Version {
     }
 }
 
-pub struct OutPathTs(PathBuf);
+pub struct OutPathTs(pub PathBuf);
 
 impl Default for OutPathTs {
     fn default() -> Self {
@@ -33,21 +27,10 @@ impl Default for OutPathTs {
     }
 }
 
-pub struct OutPathRs(PathBuf);
-
-impl Default for OutPathRs {
-    fn default() -> Self {
-        let mut path = PathBuf::from(env::var("OUT_DIR").unwrap());
-        path.push("types.rs");
-        Self(path)
-    }
-}
-
-pub fn run_transform(
+pub fn download_dts(
     Opt {
         version: Version(branch),
         out_path_ts: OutPathTs(dts_file),
-        out_path_rs: OutPathRs(rs_file),
     }: Opt,
 ) -> Result<()> {
     // setup .d.ts file
@@ -57,19 +40,6 @@ pub fn run_transform(
 
     let body = reqwest::blocking::get(url)?.text()?;
     std::fs::write(&dts_file, body)?;
-
-    let rs = dts2rs(&dts_file);
-
-    let mut writer = BufWriter::new(File::create(&rs_file)?);
-    write!(writer, "{rs}")?;
-    writer.into_inner()?;
-
-    let output = Command::new("rustfmt").arg(rs_file).output()?;
-    let status = output.status;
-    if !status.success() {
-        io::stderr().write_all(&output.stderr).unwrap();
-        panic!("failed to execute rustfmt: {status}")
-    }
 
     Ok(())
 }
